Splitting all handlers into socket and base handlers
